requirements.txt
# Core ML Libraries
torch>=2.6.0
torchvision>=0.19.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
opencv-python>=4.8.0

# Experiment Tracking
mlflow==2.8.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Data Versioning
dvc==3.8.1
dvc[s3]==3.8.1

# Monitoring
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
requests==2.31.0
matplotlib==3.8.2

